+++ OP_camera_controls.py
@@ -176,56 +176,6 @@ class STORYTOOLS_OT_camera_pan(Operator):
         
         return {'RUNNING_MODAL'}
 
-# Not really needed, already in Grease pencil tools
-class STORYTOOLS_OT_camera_rotate(Operator):
-    bl_idname = "storytools.camera_rotate"
-    bl_label = 'Rotate Camera'
-    bl_description = "Rotate camera"
-    bl_options = {'REGISTER', 'INTERNAL'}
-
-    @classmethod
-    def poll(cls, context):
-        return context.scene.camera
-
-    def invoke(self, context, event):
-        self.horizontal = event.ctrl
-        return self.execute(context)
-
-    def execute(self, context):
-        if self.horizontal: # WIP
-            ## use view
-            cam = context.scene.camera
-            aim = context.space_data.region_3d.view_rotation @ Vector((0.0, 0.0, 1.0)) # view vector
-            # aim = cam.matrix_world.to_quaternion() @ Vector((0.0, 0.0, 1.0)) # view vector
-            z_up_quat = aim.to_track_quat('Z','Y') # track Z, up Y
-            q = cam.matrix_world.to_quaternion() # store current rotation
-
-            if cam.parent:
-                q = cam.parent.matrix_world.inverted().to_quaternion() @ q
-                cam_quat = cam.parent.matrix_world.inverted().to_quaternion() @ z_up_quat
-            else:
-                cam_quat = z_up_quat
-            cam.rotation_euler = cam_quat.to_euler('XYZ')
-
-            # get diff angle (might be better way to get view axis rot diff)
-            diff_angle = q.rotation_difference(cam_quat).to_euler('ZXY').z
-            # print('diff_angle: ', math.degrees(diff_angle))
-            # set_cam_view_offset_from_angle(context, diff_angle)
-            
-            # cam.rotation_euler.z += diff_angle
-            cam.rotation_euler.rotate_axis("Z", diff_angle)
-            
-            #neg = -angle
-            #rot_mat2d = mathutils.Matrix([[math.cos(neg), -math.sin(neg)], [math.sin(neg), math.cos(neg)]])
-
-            return {"FINISHED"}
-        
-        print('Not implemented ')
-        # orientation = 'VIEW'
-        # with context.temp_override(selected_objects=[context.scene.camera]):
-        #     bpy.ops.transform.rotate('INVOKE_DEFAULT', orient_axis='Z') # orient_type='VIEW'
-        return {"FINISHED"}
-
 class STORYTOOLS_OT_attach_toggle(Operator):
     bl_idname = "storytools.attach_toggle"
     bl_label = "Attach Toggle"
@@ -401,7 +351,6 @@ def set_lockpan_km(dummy):
 classes=(
     STORYTOOLS_OT_attach_toggle,
     STORYTOOLS_OT_camera_pan,
-    STORYTOOLS_OT_camera_rotate,
     STORYTOOLS_OT_camera_depth,
     STORYTOOLS_OT_camera_lock_toggle,
     STORYTOOLS_OT_camera_key_transform,